+++ crates/viewer/re_space_view_graph/src/graph/mod.rs
@@ -1,3 +1,9 @@
+//! Provides a basic abstraction over a graph that was logged to an entity.
+
+// For now this is pretty basic, but in the future we might replace this with
+// something like `petgraph`, which will unlock more user interactions, such as
+// highlighting of neighboring nodes and edges.
+
 mod hash;
 
 use egui::{Pos2, Vec2};
@@ -9,6 +15,7 @@ use re_chunk::EntityPath;
 use re_types::components::{self, GraphType};
 
 use crate::{
+    layout::EdgeTemplate,
     ui::DrawableLabel,
     visualizers::{EdgeData, NodeData, NodeInstance},
 };
@@ -70,25 +77,10 @@ impl Node {
     }
 }
 
-pub struct Edge {
-    pub source: NodeId,
-    pub target: NodeId,
-    pub arrow: bool,
-}
-
-impl Edge {
-    pub fn id(&self) -> EdgeId {
-        EdgeId {
-            source: self.source,
-            target: self.target,
-        }
-    }
-}
-
 pub struct Graph {
     entity: EntityPath,
     nodes: Vec<Node>,
-    edges: Vec<Edge>,
+    edges: Vec<EdgeTemplate>,
     kind: GraphType,
 }
 
@@ -136,10 +128,10 @@ impl Graph {
                 }
             }
 
-            let es = data.edges.iter().map(|e| Edge {
+            let es = data.edges.iter().map(|e| EdgeTemplate {
                 source: e.source_index,
                 target: e.target_index,
-                arrow: data.graph_type == GraphType::Directed,
+                target_arrow: data.graph_type == GraphType::Directed,
             });
 
             (es.collect(), data.graph_type)
@@ -159,7 +151,7 @@ impl Graph {
         &self.nodes
     }
 
-    pub fn edges(&self) -> &[Edge] {
+    pub fn edges(&self) -> &[EdgeTemplate] {
         &self.edges
     }
