+++ crates/viewer/re_space_view_graph/src/layout/mod.rs
@@ -1,8 +1,10 @@
+mod geometry;
 mod provider;
 mod request;
 mod result;
 mod slots;
 
+pub use geometry::{EdgeGeometry, PathGeometry};
 pub use provider::ForceLayoutProvider;
-pub use request::LayoutRequest;
-pub use result::{EdgeGeometry, Layout, PathGeometry};
+pub use request::{EdgeTemplate, LayoutRequest};
+pub use result::Layout;
