+++ crates/viewer/re_space_view_graph/src/layout/result.rs
@@ -1,77 +1,11 @@
-use egui::{Pos2, Rect, Vec2};
+//! Defines the output of a layout algorithm, i.e. everything that we need to render the graph.
+
+use egui::Rect;
 use re_chunk::EntityPath;
 
 use crate::graph::{EdgeId, NodeId};
 
-#[derive(Clone, Debug)]
-pub enum PathGeometry {
-    /// A simple straight edge.
-    Line { source: Pos2, target: Pos2 },
-
-    /// Represents a cubic bezier curve.
-    ///
-    /// In the future we could probably support more complex splines.
-    CubicBezier {
-        source: Pos2,
-        target: Pos2,
-        control: [Pos2; 2],
-    },
-    // We could add other geometries, such as `Orthogonal` here too.
-}
-
-#[derive(Debug)]
-pub struct EdgeGeometry {
-    pub target_arrow: bool,
-    pub path: PathGeometry,
-}
-
-impl EdgeGeometry {
-    pub fn bounding_rect(&self) -> Rect {
-        match self.path {
-            PathGeometry::Line { source, target } => Rect::from_two_pos(source, target),
-            // TODO(grtlr): This is just a crude (upper) approximation, as the resulting bounding box can be too large.
-            PathGeometry::CubicBezier {
-                source,
-                target,
-                ref control,
-            } => Rect::from_points(&[&[source, target], control.as_slice()].concat()),
-        }
-    }
-
-    pub fn source_pos(&self) -> Pos2 {
-        match self.path {
-            PathGeometry::Line { source, .. } | PathGeometry::CubicBezier { source, .. } => source,
-        }
-    }
-
-    pub fn target_pos(&self) -> Pos2 {
-        match self.path {
-            PathGeometry::Line { target, .. } | PathGeometry::CubicBezier { target, .. } => target,
-        }
-    }
-
-    /// The direction of the edge at the source node (normalized).
-    pub fn source_arrow_direction(&self) -> Vec2 {
-        use PathGeometry::{CubicBezier, Line};
-        match self.path {
-            Line { source, target } => (source.to_vec2() - target.to_vec2()).normalized(),
-            CubicBezier {
-                source, control, ..
-            } => (control[0].to_vec2() - source.to_vec2()).normalized(),
-        }
-    }
-
-    /// The direction of the edge at the target node (normalized).
-    pub fn target_arrow_direction(&self) -> Vec2 {
-        use PathGeometry::{CubicBezier, Line};
-        match self.path {
-            Line { source, target } => (target.to_vec2() - source.to_vec2()).normalized(),
-            CubicBezier {
-                target, control, ..
-            } => (target.to_vec2() - control[1].to_vec2()).normalized(),
-        }
-    }
-}
+use super::EdgeGeometry;
 
 #[derive(Debug)]
 pub struct Layout {
