+++ crates/viewer/re_space_view_graph/src/layout/slots.rs
@@ -1,11 +1,14 @@
-//! Force-directed graph layouts assume edges to be straight lines. This
-//! function brings edges in a canonical form and finds the ones that occupy the same space.
+//! Force-directed graph layouts assume edges to be straight lines. A [`Slot`]
+//! represents the space that a single edge, _or multiple_ edges can occupy between two nodes.
+//!
+//! We achieve this by bringing edges into a canonical form via [`SlotId`], which
+//! we can then use to find duplicates.
 
 use crate::graph::NodeId;
 
 use super::request::EdgeTemplate;
 
-/// Uniquely identifies an [`EdgeSlot`] by ordering the [`NodeIds`](NodeId).
+/// Uniquely identifies an [`EdgeSlot`] by ordering the [`NodeIds`](NodeId) that make up an edge.
 #[derive(Clone, Copy, Hash, PartialEq, Eq)]
 pub struct SlotId(NodeId, NodeId);
 
@@ -19,13 +22,14 @@ impl SlotId {
     }
 }
 
+/// There are different types of [`Slots`](Slot) that are laid out differently.
 #[derive(Clone, Copy, PartialEq, Eq)]
 pub enum SlotKind {
     /// An edge slot going from `source` to `target`. Source and target represent the canonical order of the slot, as specified by [`SlotId`]
     Regular { source: NodeId, target: NodeId },
 
     /// An edge where `source == target`.
-    SelfEdge,
+    SelfEdge { node: NodeId },
 }
 
 pub struct Slot<'a> {
@@ -33,6 +37,7 @@ pub struct Slot<'a> {
     pub edges: Vec<&'a EdgeTemplate>,
 }
 
+/// Converts a list of edges into their slotted form.
 pub fn slotted_edges<'a>(
     edges: impl Iterator<Item = &'a EdgeTemplate>,
 ) -> ahash::HashMap<SlotId, Slot<'a>> {
@@ -42,7 +47,7 @@ pub fn slotted_edges<'a>(
         let id = SlotId::new(e.source, e.target);
         let slot = slots.entry(id).or_insert_with_key(|id| Slot {
             kind: match e.source == e.target {
-                true => SlotKind::SelfEdge,
+                true => SlotKind::SelfEdge{ node: e.source },
                 false => SlotKind::Regular {
                     source: id.0,
                     target: id.1,
@@ -56,5 +61,3 @@ pub fn slotted_edges<'a>(
 
     slots
 }
-
-// TODO(grtlr): Write test cases
